# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user/profile":{"get":{"operationId":"UserController_profile","summary":"","description":"This endpoint used to get current session user profile","parameters":[],"responses":{"200":{"description":"Success Response"},"400":{"description":"Bad Request Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedResponse"}}}}},"tags":["User API"],"security":[{"bearer":[]}]}},"/user/update":{"put":{"operationId":"UserController_updateUserData","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Success Response"},"400":{"description":"Bad Request Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundResponse"}}}}},"tags":["User API"],"security":[{"bearer":[]}]}},"/user/status/{id}":{"patch":{"operationId":"UserController_patchStatusUser","summary":"","description":"This endpoint used to update status user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Success Response"},"400":{"description":"Bad Request Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundResponse"}}}}},"tags":["User API"],"security":[{"bearer":[]}]}},"/user/list":{"get":{"operationId":"UserController_getUserList","summary":"","description":"This endpoint used to list user","parameters":[{"name":"search","required":false,"in":"query","example":"ID","description":"will search by id, username","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","example":1,"schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","example":10,"schema":{"type":"number"}},{"name":"orderBy","required":true,"in":"query","example":"createdAt","schema":{"type":"string"}},{"name":"orderType","required":true,"in":"query","example":"DESC","schema":{"type":"string"}}],"responses":{"200":{"description":"Success Response"},"400":{"description":"Bad Request Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnauthorizedResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundResponse"}}}}},"tags":["User API"],"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}}},"info":{"title":"Attendance NFC","description":"API for Attendance NFC","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ErrorResponse":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"Error response message"},"error":{"type":"string","example":"Error response message"}},"required":["statusCode","message","error"]},"UnauthorizedResponse":{"type":"object","properties":{"statusCode":{"type":"number","example":401},"message":{"type":"string","example":"Unauthorized"},"error":{"type":"string","example":"Unauthorized"}},"required":["statusCode","message","error"]},"UpdateUserDto":{"type":"object","properties":{"role":{"type":"string","enum":["admin","user"]},"id":{"type":"number","example":"1"},"username":{"type":"string","example":"admin"},"status":{"type":"boolean","example":"true"}},"required":["role","id","username","status"]},"NotFoundResponse":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"Not found"},"error":{"type":"string","example":"Not found"}},"required":["statusCode","message","error"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"admin"},"password":{"type":"string","example":"Your$ecr3tPass"}},"required":["username","password"]},"RegisterDto":{"type":"object","properties":{"password":{"type":"string","example":"$ecretPass"},"passwordRepeat":{"type":"string","example":"$ecretPass"},"role":{"type":"string","enum":["admin","user"]},"username":{"type":"string","example":"admin"},"status":{"type":"boolean","example":"true"}},"required":["password","passwordRepeat","role","username","status"]}}}}